Broke up migrations into smaller chunks.
