+++ alembic/versions/0022_added_audit_trigger_for_events.py
@@ -1,0 +1,72 @@
+"""
+Added audit trigger for events
+
+Revision ID: 0022
+Revises: 0021
+Create Date: 2023-11-22 22:26:17.873198
+
+"""
+import sqlalchemy as sa
+from alembic_utils.pg_trigger import PGTrigger
+
+from alembic import op
+
+# revision identifiers, used by Alembic.
+revision = "0022"
+down_revision = "0021"
+branch_labels = None
+depends_on = None
+
+
+def upgrade():
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.add_column(
+        "family_event",
+        sa.Column(
+            "created",
+            sa.DateTime(timezone=True),
+            server_default=sa.text("now()"),
+            nullable=False,
+        ),
+    )
+    op.add_column(
+        "family_event",
+        sa.Column(
+            "last_modified",
+            sa.DateTime(timezone=True),
+            server_default=sa.text("now()"),
+            nullable=False,
+        ),
+    )
+    public_family_event_update_last_modified = PGTrigger(
+        schema="public",
+        signature="update_last_modified",
+        on_entity="public.family_event",
+        is_constraint=False,
+        definition="""
+    BEFORE UPDATE ON public.family_event
+    FOR EACH ROW
+    EXECUTE PROCEDURE public.update_1_last_modified()""",
+    )
+    op.create_entity(public_family_event_update_last_modified)  # type: ignore
+
+    # ### end Alembic commands ###
+
+
+def downgrade():
+    # ### commands auto generated by Alembic - please adjust! ###
+    public_family_event_update_last_modified = PGTrigger(
+        schema="public",
+        signature="update_last_modified",
+        on_entity="public.family_event",
+        is_constraint=False,
+        definition="""
+    BEFORE UPDATE ON public.family_event
+    FOR EACH ROW
+    EXECUTE PROCEDURE public.update_1_last_modified()""",
+    )
+    op.drop_entity(public_family_event_update_last_modified)  # type: ignore
+
+    op.drop_column("family_event", "last_modified")
+    op.drop_column("family_event", "created")
+    # ### end Alembic commands ###
