+++ alembic/versions/0023_added_audit_trigger_for_family.py
@@ -1,0 +1,150 @@
+"""
+Added audit trigger for family
+
+Revision ID: 0023
+Revises: 0022
+Create Date: 2023-11-22 22:30:02.474436
+
+"""
+import sqlalchemy as sa
+from alembic_utils.pg_function import PGFunction
+from alembic_utils.pg_trigger import PGTrigger
+
+from alembic import op
+
+# revision identifiers, used by Alembic.
+revision = "0023"
+down_revision = "0022"
+branch_labels = None
+depends_on = None
+
+
+def upgrade():
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.add_column(
+        "family",
+        sa.Column(
+            "created",
+            sa.DateTime(timezone=True),
+            server_default=sa.text("now()"),
+            nullable=False,
+        ),
+    )
+    op.add_column(
+        "family",
+        sa.Column(
+            "last_modified",
+            sa.DateTime(timezone=True),
+            server_default=sa.text("now()"),
+            nullable=False,
+        ),
+    )
+    public_update_2_family_last_modified = PGFunction(
+        schema="public",
+        signature="update_2_family_last_modified()",
+        definition="""
+    RETURNS TRIGGER AS $$
+    BEGIN
+        UPDATE family
+        SET last_modified = NOW()
+        WHERE import_id = NEW.family_import_id;
+        RETURN NEW;
+    END;
+    $$ language 'plpgsql'""",
+    )
+    op.create_entity(public_update_2_family_last_modified)  # type: ignore
+
+    public_family_update_last_modified = PGTrigger(
+        schema="public",
+        signature="update_last_modified",
+        on_entity="public.family",
+        is_constraint=False,
+        definition="""
+    BEFORE UPDATE ON public.family
+    FOR EACH ROW
+    EXECUTE PROCEDURE public.update_1_last_modified()""",
+    )
+    op.create_entity(public_family_update_last_modified)  # type: ignore
+
+    public_family_document_update_family_last_modified = PGTrigger(
+        schema="public",
+        signature="update_family_last_modified",
+        on_entity="public.family_document",
+        is_constraint=False,
+        definition="""
+    BEFORE INSERT OR UPDATE OR DELETE ON public.family_document
+    FOR EACH ROW
+    EXECUTE PROCEDURE public.update_2_family_last_modified()""",
+    )
+    op.create_entity(public_family_document_update_family_last_modified)  # type: ignore
+
+    public_family_event_update_family_last_modified = PGTrigger(
+        schema="public",
+        signature="update_family_last_modified",
+        on_entity="public.family_event",
+        is_constraint=False,
+        definition="""
+    BEFORE INSERT OR UPDATE OR DELETE ON public.family_event
+    FOR EACH ROW
+    EXECUTE PROCEDURE public.update_2_family_last_modified()""",
+    )
+    op.create_entity(public_family_event_update_family_last_modified)  # type: ignore
+
+    # ### end Alembic commands ###
+
+
+def downgrade():
+    # ### commands auto generated by Alembic - please adjust! ###
+    public_family_event_update_family_last_modified = PGTrigger(
+        schema="public",
+        signature="update_family_last_modified",
+        on_entity="public.family_event",
+        is_constraint=False,
+        definition="""
+    BEFORE INSERT OR UPDATE OR DELETE ON public.family_event
+    FOR EACH ROW
+    EXECUTE PROCEDURE public.update_2_family_last_modified()""",
+    )
+    op.drop_entity(public_family_event_update_family_last_modified)  # type: ignore
+
+    public_family_document_update_family_last_modified = PGTrigger(
+        schema="public",
+        signature="update_family_last_modified",
+        on_entity="public.family_document",
+        is_constraint=False,
+        definition="""
+    BEFORE INSERT OR UPDATE OR DELETE ON public.family_document
+    FOR EACH ROW
+    EXECUTE PROCEDURE public.update_2_family_last_modified()""",
+    )
+    op.drop_entity(public_family_document_update_family_last_modified)  # type: ignore
+
+    public_family_update_last_modified = PGTrigger(
+        schema="public",
+        signature="update_last_modified",
+        on_entity="public.family",
+        is_constraint=False,
+        definition="""
+    BEFORE UPDATE ON public.family
+    FOR EACH ROW
+    EXECUTE PROCEDURE public.update_1_last_modified()""",
+    )
+    op.drop_entity(public_family_update_last_modified)  # type: ignore
+
+    public_update_2_family_last_modified = PGFunction(
+        schema="public",
+        signature="update_2_family_last_modified()",
+        definition="""
+    RETURNS TRIGGER AS $$
+    BEGIN
+        UPDATE family
+        SET last_modified = NOW()
+        WHERE import_id = NEW.family_import_id;
+        RETURN NEW;
+    END;
+    $$ language 'plpgsql'""",
+    )
+    op.drop_entity(public_update_2_family_last_modified)  # type: ignore
+    op.drop_column("family", "last_modified")
+    op.drop_column("family", "created")
+    # ### end Alembic commands ###
