+++ app/db/models/law_policy/collection.py
@@ -1,7 +1,7 @@
 import sqlalchemy as sa
-from app.db.models.law_policy import Family
-from app.db.models.app import Organisation
 
+from app.db.models.app import Organisation
+from app.db.models.law_policy import Family
 from app.db.session import Base
